+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+What's changed since v0.2.0:
+
+- New features:
+  - Added support for passing through rule annotations. [#29](https://github.com/microsoft/PSRule.Monitor/issues/29)
+  - Added convention to support ingestion in a pipeline. [#46](https://github.com/microsoft/PSRule.Monitor/issues/46)
+    - To use this feature include the `Monitor.LogAnalytics.Import` convention.
+- General improvements:
+  - Ignore null or empty field and data properties. [#44](https://github.com/microsoft/PSRule.Monitor/issues/44)
+  - Added `CorrelationId` to ingested events. [#45](https://github.com/microsoft/PSRule.Monitor/issues/44)
+  - Added configuration of workspace parameter through environment variable. [#32](https://github.com/microsoft/PSRule.Monitor/issues/32)
+    - Set `PSRULE_CONFIGURATION_MONITOR_WORKSPACE_ID` to configure the workspace id.
+    - Set `PSRULE_CONFIGURATION_MONITOR_WORKSPACE_KEY` to configure the shared key.
+
 ## v0.2.0
 
 What's changed since v0.1.0: